+++ openbbs/cli.py
@@ -2,7 +2,7 @@
 
 import argparse
 
-from asciichan import (__doc__, __version__)
+from openbbs import (__doc__, __version__)
 
 
 class CustomHelp(argparse.HelpFormatter):
@@ -41,7 +41,7 @@ docs.add_argument(
     "-v",
     "--version",
     action="version",
-    version="Asciichan Server, Version %s." % __version__,
+    version="OpenBBS Server, Version %s." % __version__,
     help="Display the currently installed version and exit."
 )
 
@@ -58,6 +58,6 @@ server_opts.add_argument(
     "--config",
     metavar="XX",
     default="./config.ini",
-    help="Specify the path of the config.ini file that the Asciichan\nserver "
+    help="Specify the path of the config.ini file that the OpenBBS\nserver "
     "should read from."
 )