+++ tests/test_core.py
@@ -4,18 +4,16 @@ import threading
 import time
 import unittest
 
-from tests.dummy_objects import DummyConfig
-
-from asciichan.core import spawn_server
+from openbbs.config import load_config
+from openbbs.core import spawn_server
 
 logging.disable(logging.CRITICAL)
 
 
 class ServerSpawnTest(unittest.TestCase):
     def test_create_server(self):
-        fakeconfig = DummyConfig()
-        threading.Thread(target=spawn_server, args=(fakeconfig.fakeget,
-                                                    True)).start()
+        config = load_config("inexistent.conf")
+        threading.Thread(target=spawn_server, args=(config, True)).start()
         # Connection occurs before binding if the test doesn't sleep. 
         time.sleep(1)
         test_sock = socket.socket(socket.AF_INET, socket.SOCK_STREAM)
@@ -25,6 +23,7 @@ class ServerSpawnTest(unittest.TestCase):
         test_sock.close()
 
     def test_failure_to_bind(self):
-        fakeconfig = DummyConfig(server={"host": "999.999.999"})
+        config = load_config("inexistent.conf")
+        config["host"] = "999.999.999.999"
         with self.assertRaises(SystemExit):
-            spawn_server(fakeconfig.fakeget, True)
+            spawn_server(config, True)